+++ kartothek/io/testing/gc.py
@@ -18,7 +18,7 @@ def _test_gc(uuid, store_factory, garbage_collect_callable):
     keys_before = set(store.keys())
 
     # Add a non-tracked table file
-    store.put("{}/core/trash.parquet".format(uuid), b"trash")
+    store.put("{}/table/trash.parquet".format(uuid), b"trash")
 
     # Add a non-tracked index file
     store.put("{}/indices/trash.parquet".format(uuid), b"trash")